refactor: integrate API with Juno
